Productservice validations

+++ Controllers/ProductController.cs
@@ -26,50 +26,66 @@ public class ProductController : ControllerBase
     {
         List<Product> products = await _productService.GetProducts();
         if (products.Count == 0)
-            return NoContent();
-        else
-            return products;
+            return NotFound("No Products exist in database");
+
+        return products;
     }
 
     [HttpGet("{id}")]
     public async Task<ActionResult<Product>> GetProduct(int id)
     {
+        if (id == null)
+            return BadRequest("productId cannot be null");
         Console.WriteLine("--- debug ---- product.Id: " + id);
 
         Product product = await _productService.GetProduct(id);
         if (product == null)
-            return NotFound();
+            return NotFound("No product found for this id");
         return product;
     }
 
     [HttpGet("search/{productName}")]
-    public async Task<List<Product>> SearchProducts(String productName)
+    public async Task<ActionResult<List<Product>>> SearchProducts(String productName)
     {
+        if (productName == null)
+            return BadRequest("Product name cannot be null");
         Console.WriteLine("--- debug ---- product.name: " + productName);
 
         List<Product> searchedProducts = await _productService.SearchProducts(productName);
 
         if (searchedProducts.Count == 0)
-            return new List<Product>();
+            return NotFound("No products found for this product name");
+
         return searchedProducts;
     }
 
     [HttpGet("search-cat/{categoryName}")]
-    public async Task<List<Product>> SearchCategoryProducts(String categoryName)
+    public async Task<ActionResult<List<Product>>> SearchCategoryProducts(String categoryName)
     {
+        if (categoryName == null)
+            return BadRequest("Category name param cannot be null");
+
         Console.WriteLine("--- debug ---- product.name: " + categoryName);
 
         List<Product> searchedProducts = await _productService.SearchCategoryProducts(categoryName);
 
         if (searchedProducts.Count == 0)
-            return new List<Product>();
+            return NotFound("No products found for this category name");
+       
         return searchedProducts;
     }
 
     [HttpPost]
     public async Task<ActionResult<bool>> AddProduct(Product product)
     {
+        if (product.Id == null || product.Name == null || product.Price == null
+            || product.Quantity == null || product.Description == null)
+            return BadRequest("One of the body params is null");
+
         bool result = await _productService.AddProduct(product);
+
+        if (result == false)
+            return NotFound("Product with same id or name exists");
 
         return result;
     }
@@ -77,16 +93,29 @@ public class ProductController : ControllerBase
     [HttpPut]
     public async Task<ActionResult<bool>> EditProduct (Product product)
     {
+        if (product.Id == null || product.Name == null || product.Price == null
+            || product.Quantity == null || product.Description == null)
+            return BadRequest("One of the body params is null");
+
         bool result = await _productService.EditProduct(product);
+
         if (result == false)
-            return NotFound();
+            return NotFound("Product does not exist so cannot update");
+
         return result;
     }
 
     [HttpDelete("{Id}")]
     public async Task<ActionResult<bool>> DeleteProduct(int Id)
     {
+        if (Id == null)
+            return BadRequest("ProductId cannot be null");
+
         bool result = await _productService.DeleteProductById(Id);
+
+        if (result == false)
+            return NotFound("Cannot delete product because it does not exist");
+
         return result;
     }
 }
